Clean Up and Additions to the files
Deleted a lot of the files that weren't being used for the current
production of the game. Added a folder for versions of the code so
whenever we get a portion of the code functioning properly, we
Copy+Paste the code into a new Action Script File so we can see the most
recent working code.